Update pyo3 to 0.22
This removes the `Clone` impl from `Py<T>` which we can work around by
keeping the GIL a bit longer in `log`. I believe that this has little to
no implication (or maybe a positive one actually due to the switch to
`clone_ref`).

However, this means that we have to remove `Clone` impl from `Logger`.

All of this is still compatible with `0.21` but still a breaking change.